IoT.js string size optimization 32 -> 16
1. Capacity and size fields of IoT.js string were changed from
   unsigned int to unsigned short int.
2. Two bytes was saved in structure size.
3. Additional asserts were introduced to check for max size.

Verification:
 $ tools/precommit.py

All automated tests at 64-bits Linux PC are passing.

Please execute test_fs_readfileSync_null_string.js manually and check
if both test/resources/null.bin and /tmp/null.bin are identical.

IoT.js-DCO-1.0-Signed-off-by: Piotr Marcinkiewicz p.marcinkiew@samsung.com

+++ src/iotjs_string.h
@@ -18,8 +18,8 @@
 
 
 typedef struct {
-  unsigned size;
-  unsigned cap;
+  unsigned short size;
+  unsigned short cap;
   char* data;
 } IOTJS_VALIDATED_STRUCT(iotjs_string_t);
